+++ src/crypto/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod encryption;
+mod master;
+pub(crate) mod signing;
+
+pub use master::{Key, KeyError};
